requirements.txt

flask>=3.0,<4.0

PyMuPDF>=1.24.0


sentence-transformers>=2.7.0


faiss-cpu>=1.8.0


transformers>=4.40.0
torch>=2.2.0            
sentencepiece>=0.2.0   
protobuf>=3.20.0       


mysql-connector-python>=8.4.0
